Long polls now time out
Before this, all new pages started their long poll threads that kept
polling forever. Now these polls time out, and it is the responsibility of the JS running on the page to restart it, provided it is still open.

Also, added a whole lot of debug messages to see which polling threads are still running.